fix(app): host native MarkdownSource in View
Inline MarkdownSource wrapped children in Text. Plugins that draw with
SvgXml nest a native view there, which iOS and Android reject. Native
has no DOM selection, so both inline and display use View.
